qcom: ssr: Allow taking system ramdumps when system_debug is set
Add a new system_debug subsystem property. This is set as
"reset" by default. When this parameter is listed as "set"
for a subsystem and a wdog bite happens, the SSR framework
goes ahead and triggers a kernel panic.
This feature is strictly intented only for debugging cases
where a system ramdump is more desirable than subsystem
ramdumps.

CRs-Fixed: 746414
Change-Id: Ie8f364fe0f9fd8f417f080135ab497e816f7fcf4
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Peter Nyilas <peter.nyilas@byyd-tech.com>