net: msm_rmnet: Add an IOCTL definition to deregister a net device
Platform drivers may be registered on driver startup for multiple
channels which may cause probe and remove events to be called
corresponding to driver events. It is possible that not all
channels need to be enumerated for normal device operation. The
unused channels also add unnecessary delay during sub system
restart.

Add an extended IOCTL which provides user space an option to
unregister a platform driver.

CRs-Fixed: 738039
Change-Id: Ifa6284563551ecdd297ae457ea2b35027f60844e
Acked-by: Nagarjuna Chaganti <nchagant@qti.qualcomm.com>
Signed-off-by: Ravinder Konka <rkonka@codeaurora.org>
Signed-off-by: Subash Abhinov Kasiviswanathan <subashab@codeaurora.org>
Signed-off-by: Peter Nyilas <peter.nyilas@byyd-tech.com>